Fix tests. Rename Joint Surveys
