+++ geoapps/inversion/__init__.py
@@ -40,9 +40,9 @@ DRIVER_MAP = {
         "geoapps.inversion.electricals.induced_polarization.pseudo_three_dimensions.driver",
         "InducedPolarizationPseudo3DDriver",
     ),
-    "joint single property": (
-        "geoapps.inversion.joint.single_property.driver",
-        "JointSingleDriver",
+    "joint surveys": (
+        "geoapps.inversion.joint.joint_surveys.driver",
+        "JointSurveyDriver",
     ),
     "tdem": (
         "geoapps.inversion.airborne_electromagnetics.time_domain.driver",
